+++ src/dodal/devices/beamlines/i17/i17_apple2.py
@@ -1,11 +1,12 @@
 from dodal.devices.insertion_device import (
     Apple2,
     Apple2Controller,
+    Apple2PhasesVal,
     Apple2Val,
     EnergyMotorLookup,
     Pol,
+    UndulatorPhaseAxes,
 )
-from dodal.devices.insertion_device.undulator import Apple2PhasesVal, UndulatorPhaseAxes
 
 ROW_PHASE_MOTOR_TOLERANCE = 0.004
 MAXIMUM_ROW_PHASE_MOTOR_POSITION = 24.0
