+++ src/dodal/devices/insertion_device/__init__.py
@@ -1,11 +1,13 @@
-from .apple2 import Apple2, Apple2Val
+from .access_control import UndulatorAccessControl
 from .apple2_controller import (
     MAXIMUM_GAP_MOTOR_POSITION,
     MAXIMUM_MOVE_TIME,
     MAXIMUM_ROW_PHASE_MOTOR_POSITION,
+    Apple2,
     Apple2Controller,
     Apple2EnforceLHMoveController,
     Apple2Type,
+    Apple2Val,
     EnergyMotorConvertor,
 )
 from .apple_knot_controller import AppleKnotController, AppleKnotPathFinder
@@ -17,16 +19,29 @@ from .energy_motor_lookup import (
     StaticPolynomialEnergyMotorLookup,
 )
 from .enum import Pol, UndulatorGateStatus
+from .gap import UndulatorGap
 from .lookup_table_models import (
     EnergyCoverage,
     EnergyCoverageEntry,
     LookupTable,
     LookupTableColumnConfig,
+    Source,
     convert_csv_to_lookup,
+)
+from .phase_axes import (
+    Apple2LockedPhasesVal,
+    Apple2PhasesVal,
+    Apple2PhaseValType,
+    PhaseAxesType,
+    UndulatorJawPhase,
+    UndulatorLockedPhaseAxes,
+    UndulatorPhaseAxes,
+    UndulatorPhaseMotor,
 )
 from .polarisation import InsertionDevicePolarisation
 
 __all__ = [
+    "UndulatorAccessControl",
     "Apple2",
     "Apple2Type",
     "Apple2Controller",
@@ -36,6 +51,7 @@ __all__ = [
     "Apple2Val",
     "MAXIMUM_MOVE_TIME",
     "LookupTable",
+    "Source",
     "LookupTableColumnConfig",
     "convert_csv_to_lookup",
     "InsertionDeviceEnergy",
@@ -44,6 +60,7 @@ __all__ = [
     "EnergyCoverage",
     "Pol",
     "UndulatorGateStatus",
+    "UndulatorGap",
     "EnergyMotorLookup",
     "ConfigServerEnergyMotorLookup",
     "EpicsPolynomialEnergyMotorLookup",
@@ -52,4 +69,12 @@ __all__ = [
     "MAXIMUM_ROW_PHASE_MOTOR_POSITION",
     "MAXIMUM_GAP_MOTOR_POSITION",
     "StaticPolynomialEnergyMotorLookup",
+    "UndulatorJawPhase",
+    "Apple2LockedPhasesVal",
+    "Apple2PhasesVal",
+    "Apple2PhaseValType",
+    "PhaseAxesType",
+    "UndulatorLockedPhaseAxes",
+    "UndulatorPhaseAxes",
+    "UndulatorPhaseMotor",
 ]
