+++ src/dodal/devices/insertion_device/phase_axes.py
@@ -8,10 +8,10 @@ from bluesky.protocols import Movable
 from ophyd_async.core import AsyncStatus, Reference, StandardReadable
 from ophyd_async.epics.core import epics_signal_r, epics_signal_rw
 
-from dodal.devices.insertion_device.undulator.access_control import (
+from dodal.devices.insertion_device.access_control import (
     UndulatorAccessControl,
 )
-from dodal.devices.insertion_device.undulator.utils import (
+from dodal.devices.insertion_device.utils import (
     MotorStringSetpoint,
     UndulatorCoordinatableMove,
     UnstoppableMotorMoveLogic,
