--- src/dodal/devices/insertion_device/undulator/__init__.py
@@ -1,19 +1,0 @@
-from .access_control import UndulatorAccessControl
-from .gap import UndulatorGap
-from .phase_axes import (
-    Apple2LockedPhasesVal,
-    Apple2PhasesVal,
-    UndulatorJawPhase,
-    UndulatorLockedPhaseAxes,
-    UndulatorPhaseAxes,
-)
-
-__all__ = [
-    "UndulatorGap",
-    "UndulatorAccessControl",
-    "Apple2LockedPhasesVal",
-    "Apple2PhasesVal",
-    "UndulatorJawPhase",
-    "UndulatorLockedPhaseAxes",
-    "UndulatorPhaseAxes",
-]
